+++ apps/cli/src/domain/environment.ts
@@ -85,8 +85,7 @@ type CloudAccountNotInitializedIssue = {
 };
 
 type EnvironmentInitIssue =
-  | CloudAccountNotInitializedIssue
-  | MissingRemoteBackendIssue;
+  CloudAccountNotInitializedIssue | MissingRemoteBackendIssue;
 
 type MissingRemoteBackendIssue = {
   type: "MISSING_REMOTE_BACKEND";