+++ actions/pr-comment/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.3.5 (2026-07-28)
+
+### 🧱 Updated Dependencies
+
+- Updated @pagopa/eslint-config to 6.2.2
+
 ## 0.3.3 (2026-07-02)
 
 ### 🧱 Updated Dependencies